Add knobs storybook addon
