+++ .storybook/addons.js
@@ -1,0 +1,2 @@
+import '@kadira/storybook/addons';
+import '@kadira/storybook-addon-knobs/register';
