+++ .storybook/config.js
@@ -1,7 +1,12 @@
+/* eslint global-require: "off" */
 import { configure } from '@kadira/storybook';
 
+const reqireStories = require.context('../src', true, /.stories.js$/);
+
 function loadStories() {
-  require('../stories/index.js');
+  require('./decorators');
+
+  reqireStories.keys().forEach(filename => reqireStories(filename));
 }
 
 configure(loadStories, module);
