+++ package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@kadira/storybook": "^2.20.1",
+    "@kadira/storybook-addon-knobs": "^1.3.0",
     "enzyme": "^2.4.1",
     "react": "^15.3.2",
     "react-dom": "^15.3.2"
